Refactor io_data.h and cpp into separate xdata h and cpp files
