+++ companion/src/firmwares/heli_data.h
@@ -1,0 +1,44 @@
+/*
+ * Copyright (C) OpenTX
+ *
+ * Based on code named
+ *   th9x - http://code.google.com/p/th9x
+ *   er9x - http://code.google.com/p/er9x
+ *   gruvin9x - http://code.google.com/p/gruvin9x
+ *
+ * License GPLv2: http://www.gnu.org/licenses/gpl-2.0.html
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License version 2 as
+ * published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ */
+
+#pragma once
+
+#include "rawsource.h"
+
+#include <QtCore>
+
+class SwashRingData {
+  Q_DECLARE_TR_FUNCTIONS(SwashRingData)
+
+  public:
+    SwashRingData() { clear(); }
+
+    int elevatorWeight;
+    int aileronWeight;
+    int collectiveWeight;
+    unsigned int  type;
+    RawSource collectiveSource;
+    RawSource aileronSource;
+    RawSource elevatorSource;
+    unsigned int  value;
+
+    void clear() { memset(reinterpret_cast<void *>(this), 0, sizeof(SwashRingData)); }
+};
+
